preparing to refactor spack solver setup
In anticipation of being able to run one or more custom models (for testing
or future production) we need to untangle the spack solver setup from the single
asp.py, and have a means to automatically discover (and then ask to use) one or
more setups. In this case, a setup includes a full class that expects a driver
to generate facts, and associated with that class is one or more logic programs
to run. To keep things tidy, the logic programs are now stored in their own logic
directory. Since this PR is just refactoring the current (and we only have one)
there is no argument exposed in spack.cmd.solve. However, when I add more models
(e.g., for symbols) there will be this option.
