+++ lib/spack/spack/paths.py
@@ -38,6 +38,7 @@ platform_path         = os.path.join(module_path, 'platforms')
 compilers_path        = os.path.join(module_path, "compilers")
 build_systems_path    = os.path.join(module_path, 'build_systems')
 operating_system_path = os.path.join(module_path, 'operating_systems')
+models_path           = os.path.join(module_path, 'solver', 'models')
 test_path             = os.path.join(module_path, "test")
 hooks_path            = os.path.join(module_path, "hooks")
 var_path              = os.path.join(prefix, "var", "spack")
