+++ lib/spack/spack/solver/logic/display.lp
